Show loading spinner for runner while the project is loading (#892)
* Show loading spinner for runner while the project is loading

* move loading out of IframeDisplay

* show loading on re-run
