+++ spx-gui/src/components/project/runner/IframeDisplay.vue
@@ -1,12 +1,10 @@
 <template>
-  <div class="iframe-container">
-    <iframe ref="iframe" class="iframe" frameborder="0" src="about:blank" />
-    <UILoading :visible="loading" cover />
-  </div>
+  <iframe ref="iframe" class="iframe" frameborder="0" src="about:blank" />
 </template>
 <script setup lang="ts">
 const emit = defineEmits<{
   console: [type: 'log' | 'warn', args: unknown[]]
+  loaded: []
 }>()
 
 interface IframeWindow extends Window {
@@ -19,16 +17,16 @@ import rawRunnerHtml from '@/assets/ispx/runner.html?raw'
 import wasmExecUrl from '@/assets/wasm_exec.js?url'
 import wasmUrl from '@/assets/ispx/main.wasm?url'
 import { watch } from 'vue'
-import { UILoading } from '@/components/ui'
 
-const { zipData } = defineProps<{ zipData: ArrayBuffer | Uint8Array }>()
+const props = defineProps<{ zipData: ArrayBuffer | Uint8Array }>()
 
 const iframe = ref<HTMLIFrameElement>()
 
-const loading = ref(true)
-
 watch(iframe, () => {
-  const iframeWindow = iframe.value?.contentWindow as IframeWindow | null | undefined
+  if (!iframe.value) {
+    return
+  }
+  const iframeWindow = iframe.value.contentWindow as IframeWindow | null
   if (!iframeWindow) {
     return
   }
@@ -39,11 +37,11 @@ watch(iframe, () => {
   iframeWindow.document.write(runnerHtml) // This resets the iframe's content, including its window object
 
   iframeWindow.addEventListener('wasmReady', () => {
-    iframeWindow.startWithZipBuffer(zipData)
+    iframeWindow.startWithZipBuffer(props.zipData)
     const canvas = iframeWindow.document.querySelector('canvas')
     if (canvas == null) throw new Error('canvas expected in iframe')
     canvas.focus() // focus to canvas by default, so the user can interact with the game immediately
-    loading.value = false
+    emit('loaded')
   })
   iframeWindow.console.log = function (...args: unknown[]) {
     // eslint-disable-next-line no-console
@@ -58,13 +56,6 @@ watch(iframe, () => {
 </script>
 
 <style scoped lang="scss">
-.iframe-container {
-  position: relative;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 .iframe {
   width: 100%;
   height: 100%;
